yolo-service/requirements.txt
# YOLO Service Dependencies
# ==========================
# FastAPI for REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Ultralytics YOLO (minimum 8.4.0 for platform integration)
ultralytics>=8.4.0

# Image processing
Pillow>=10.0.0
python-dotenv>=1.0.0

# Request validation
pydantic>=2.5.0

# Async HTTP client (for health checks)
httpx>=0.26.0
